first commit changes to User.java , added endpoint /all that retrieves all Users

+++ src/main/java/com/linkedin/database/repo/UserRepository.java
@@ -2,8 +2,11 @@ package com.linkedin.database.repo;
 
 import com.linkedin.database.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
 }
